[headers] Move files to common/integration_testing/src/public/
Move all files from
//common/integration_testing/src/google_smart_card_integration_testing_common/
to //common/integration_testing/src/public/.

It's a pure refactoring commit that gets rid of unnecessarily long file
names. It's possible now that we include these headers by their
absolute path, as tracked by #885.
