+++ common/integration_testing/src/entry_point_emscripten.cc
@@ -32,7 +32,7 @@
 #include "common/cpp/src/public/messaging/typed_message_router.h"
 #include "common/cpp/src/public/value.h"
 #include "common/cpp/src/public/value_emscripten_val_conversion.h"
-#include "common/integration_testing/src/google_smart_card_integration_testing/integration_test_service.h"
+#include "common/integration_testing/src/public/integration_test_service.h"
 
 namespace google_smart_card {
