+++ common/integration_testing/src/public/integration-test-controller.js
