+++ common/integration_testing/src/public/integration_test_helper.h
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef GOOGLE_SMART_CARD_INTEGRATION_TESTING_INTEGRATION_TEST_HELPER_H_
-#define GOOGLE_SMART_CARD_INTEGRATION_TESTING_INTEGRATION_TEST_HELPER_H_
+#ifndef GOOGLE_SMART_CARD_COMMON_INTEGRATION_TESTING_SRC_PUBLIC_INTEGRATION_TEST_HELPER_H_
+#define GOOGLE_SMART_CARD_COMMON_INTEGRATION_TESTING_SRC_PUBLIC_INTEGRATION_TEST_HELPER_H_
 
 #include <functional>
 #include <string>
@@ -59,4 +59,4 @@ class IntegrationTestHelper {
 
 }  // namespace google_smart_card
 
-#endif  // GOOGLE_SMART_CARD_INTEGRATION_TESTING_INTEGRATION_TEST_HELPER_H_
+#endif  // GOOGLE_SMART_CARD_COMMON_INTEGRATION_TESTING_SRC_PUBLIC_INTEGRATION_TEST_HELPER_H_
