+++ example_cpp_smart_card_client_app/src/chrome_certificate_provider/api_bridge_integration_test_helper.cc
@@ -29,8 +29,8 @@
 #include "common/cpp/src/public/unique_ptr_utils.h"
 #include "common/cpp/src/public/value.h"
 #include "common/cpp/src/public/value_conversion.h"
-#include "common/integration_testing/src/google_smart_card_integration_testing/integration_test_helper.h"
-#include "common/integration_testing/src/google_smart_card_integration_testing/integration_test_service.h"
+#include "common/integration_testing/src/public/integration_test_helper.h"
+#include "common/integration_testing/src/public/integration_test_service.h"
 
 #include "chrome_certificate_provider/api_bridge.h"
 #include "chrome_certificate_provider/types.h"
